+++ packages/core/src/Aesthetic.ts
@@ -21,7 +21,6 @@ import {
 // contextual themeing
 // theme specific `style` elements
 // keyframes name -> hash aliasing
-// css variables within a class
 
 function createRenderer(): Renderer {
   return global.AESTHETIC_SSR || new ClientRenderer();